+++ lib/adaptive_core.dart
@@ -6,4 +6,5 @@ library adaptive_core;
 
 export 'src/adaptive_core.dart';
 export 'src/adaptive_user.dart';
+export 'src/adaptive_user_grade.dart';
 export 'src/adaptive_exception.dart';